Added models, DbContext, DbSets and Fluent API.

+++ KappaShop/Models/Design/Position.cs
@@ -10,15 +10,15 @@
         /// </summary>
         /// <param name="x">Represents the position on the x axis on a 2D space.</param>
         /// <param name="y">Represents the position on the y axis on a 2D space.</param>
-        public Position(float x, float y) => (X, Y) = (x, y);
+        public Position(double x, double y) => (X, Y) = (x, y);
 
         /// <summary>
         /// Represents the position on the x axis on a 2D space.
         /// </summary>
-        public float X { get; set; }
+        public double X { get; set; }
         /// <summary>
         /// Represents the position on the y axis on a 2D space.
         /// </summary>
-        public float Y { get; set; }
+        public double Y { get; set; }
     }
 }
